requirements.txt
asgiref==3.8.1
asttokens==2.4.1
biopython==1.84
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.6
decorator==5.1.1
dj-database-url==2.3.0
Django==5.1.1
django-crispy-forms==2.3
executing==2.1.0
gunicorn==23.0.0
h11==0.14.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==307
pyzmq==26.2.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.32.1
wcwidth==0.2.13
whitenoise==6.8.2
